Struct libimagstorestdhook::vcs::git::update::UpdateHook

pub struct UpdateHook {
    // some fields omitted
}

The UpdateHook type

Represents a hook which is executed whenever a entry in the store is updated (written to disk).

Time of execution

This hook is executed after the store operation succeeded, so after the file is written to disk.

Methods

impl UpdateHook

fn new(storepath: PathBuf, p: HookPosition) -> UpdateHook